requirements.txt
fastapi>=0.111.0
langsmith>=0.1.0
uvicorn[standard]>=0.29.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain-anthropic>=0.3.0
langchain-core>=0.2.0
google-search-results>=2.4.2
pydantic>=2.7.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
